9-item audit remediation (critical->low) plus a 4-lens adversarial red-team
that found and fixed 4 further defects. Each fix has a load-bearing regression
test; full host suite (15 engine configs + 2 compiler-emit) green under
ASan/UBSan; ctest 100%.

Known follow-up (tracked separately): AVR reads layer descriptors without
pgm_read while emitting them PROGMEM -> garbage on real AVR.

+++ src/arch/mnv_bnn.c
@@ -53,12 +53,14 @@
  * @param len     Number of binary values (in_sz).
  * @return Signed accumulator in range [-len, +len].
  */
-static int16_t bnn_dot_bits(const uint8_t *w, uint16_t w_off,
+static int16_t bnn_dot_bits(const uint8_t *w, uint32_t w_off,
                             const uint8_t *a, uint16_t len)
 {
     int16_t acc = 0;
     for (uint16_t i = 0; i < len; i++) {
-        uint16_t wi = (uint16_t)(w_off + i);
+        /* uint32: a single layer with in_sz*out_sz >= 65536 bits pushes the
+         * neuron bit offset past uint16 (would wrap and read wrong weights). */
+        uint32_t wi = w_off + (uint32_t)i;
         uint8_t  wb = (uint8_t)((w[wi >> 3] >> (wi & 7u)) & 1u);
         uint8_t  ab = (uint8_t)((a[i  >> 3] >> (i  & 7u)) & 1u);
         /* xor==0 -> agree -> +1 ; xor==1 -> disagree -> -1 (branchless) */
@@ -118,13 +120,13 @@ mnv_status_t mnv_bnn_forward(mnv_ctx_t          *ctx,
      * the *ciphertext* pointer by the same amount each step or layer >0
      * decrypts the wrong bytes. */
     const uint8_t *ct     = model->encrypted_weights;
-    uint16_t       ct_off = 0U;
+    uint32_t       ct_off = 0U;
 
     for (uint8_t layer = 0; layer < model->num_layers; layer++) {
         const mnv_layer_desc_t *ld = &model->layers[layer];
         uint16_t in_sz    = ld->input_size;
         uint16_t out_sz   = ld->output_size;
-        uint16_t w_bytes  = (in_sz * out_sz + 7u) / 8u;
+        uint32_t w_bytes  = ((uint32_t)in_sz * (uint32_t)out_sz + 7u) / 8u;
 
         /* Decrypt packed weights */
         mnv_chacha20_decrypt(chacha, ct + ct_off,
@@ -145,7 +147,7 @@ mnv_status_t mnv_bnn_forward(mnv_ctx_t          *ctx,
         for (uint16_t n = 0; n < out_sz; n++) {
             /* Weights for neuron n start at bit n*in_sz in the packed buffer. */
             int16_t acc = bnn_dot_bits((const uint8_t *)ctx->weight_scratch,
-                                       (uint16_t)(n * in_sz), packed_src, in_sz);
+                                       (uint32_t)n * (uint32_t)in_sz, packed_src, in_sz);
 
             /* Scale to Q8: acc ∈ [-in_sz, +in_sz], map to [-127, +127] */
             int16_t scaled = (int16_t)((int32_t)acc * 127 / (int16_t)in_sz);
