+++ src/security/mnv_blake2s.h
@@ -20,13 +20,13 @@ typedef struct {
 void         mnv_blake2s_init(mnv_blake2s_ctx_t *ctx,
                               const uint8_t *key, uint8_t keylen);
 void         mnv_blake2s_update(mnv_blake2s_ctx_t *ctx,
-                                const uint8_t *data, uint16_t len);
+                                const uint8_t *data, uint32_t len);
 void         mnv_blake2s_final(mnv_blake2s_ctx_t *ctx, uint8_t *digest);
 void         mnv_blake2s_mac(const uint8_t *key,  uint8_t  keylen,
-                             const uint8_t *data, uint16_t datalen,
+                             const uint8_t *data, uint32_t datalen,
                              uint8_t *digest);
 mnv_status_t mnv_blake2s_verify(const uint8_t *key,  uint8_t  keylen,
-                                const uint8_t *data, uint16_t datalen,
+                                const uint8_t *data, uint32_t datalen,
                                 const uint8_t *expected_mac);
 
 #endif /* MNV_BLAKE2S_H */
